fix: Fix build error
