+++ cxx-parser/src/cxx_parser_ext.ts
@@ -30,17 +30,17 @@ ParseResult.prototype.findClazz = function (
   const clazzNameNamespace = clazzName.getNamespace();
   const trimNamespacesName = clazzName.trimNamespace();
 
-  for (const node of this.nodes) {
-    let cxxFile = node as CXXFile;
+  for (const f of this.nodes) {
+    let cxxFile = f as CXXFile;
     for (const node of cxxFile.nodes) {
       if (node.__TYPE == CXXTYPE.Clazz) {
         const clazz = node;
 
         if (clazz.name == trimNamespacesName) {
           if (
-            clazz.getNamespacesInString() == clazzNameNamespace ||
+            clazz.namespace == clazzNameNamespace ||
             clazzNameNamespace == "" ||
-            clazz.getNamespacesInString().includes(clazzNameNamespace)
+            clazz.namespace.includes(clazzNameNamespace)
           ) {
             return clazz as Clazz;
           }
@@ -66,17 +66,17 @@ ParseResult.prototype.findEnumz = function (
 
   const trim_namespaces_name = enumzName.trimNamespace();
 
-  for (const n of this.nodes) {
-    let cxx_file = n as CXXFile;
-    for (const node of cxx_file.nodes) {
+  for (const f of this.nodes) {
+    let cxxFile = f as CXXFile;
+    for (const node of cxxFile.nodes) {
       if (node.__TYPE == CXXTYPE.Enumz) {
         const enumz = node as Enumz;
 
         if (enumz.name === trim_namespaces_name) {
           if (
-            enumz.getNamespacesInString() == namespace_string ||
+            enumz.namespace == namespace_string ||
             namespace_string == "" ||
-            enumz.getNamespacesInString().includes(namespace_string)
+            enumz.namespace.includes(namespace_string)
           ) {
             return enumz;
           }
